requirements.txt
django>=3.2,<3.3
gunicorn
psycopg2-binary
django-dotenv
pyspellchecker